+++ packages/stream_chat_flutter/CHANGELOG.md
@@ -2,8 +2,8 @@
 
 🐞 Fixed
 
-- Fixed audio tone bleeding into recorded voice message when playing custom feedback sound on recording
-  start.
+- Fixed `StreamChannelAvatar` crashing with `RangeError` when user/channel name is empty.
+- Fixed audio tone bleeding into recorded voice message when playing custom feedback sound on recording start.
 
 ## 9.22.0